+++ src/Exception/AccountNotActiveException.php
@@ -16,4 +16,15 @@ use Cake\Core\Exception\Exception;
 class AccountNotActiveException extends Exception
 {
     protected $_messageTemplate = '/a/validate/%s/%s';
+
+    /**
+     * AccountNotActiveException constructor.
+     * @param array|string $message message
+     * @param int $code code
+     * @param null $previous previous
+     */
+    public function __construct($message, $code = 500, $previous = null)
+    {
+        parent::__construct($message, $code, $previous);
+    }
 }