+++ src/Shell/UsersShell.php
@@ -56,15 +56,33 @@ class UsersShell extends Shell
     {
         $parser = parent::getOptionParser();
         $parser->setDescription(__d('CakeDC/Users', 'Utilities for CakeDC Users Plugin'))
-            ->addSubcommand('activateUser')->setDescription(__d('CakeDC/Users', 'Activate an specific user'))
-            ->addSubcommand('addSuperuser')->setDescription(__d('CakeDC/Users', 'Add a new superadmin user for testing purposes'))
-            ->addSubcommand('addUser')->setDescription(__d('CakeDC/Users', 'Add a new user'))
-            ->addSubcommand('changeRole')->setDescription(__d('CakeDC/Users', 'Change the role for an specific user'))
-            ->addSubcommand('deactivateUser')->setDescription(__d('CakeDC/Users', 'Deactivate an specific user'))
-            ->addSubcommand('deleteUser')->setDescription(__d('CakeDC/Users', 'Delete an specific user'))
-            ->addSubcommand('passwordEmail')->setDescription(__d('CakeDC/Users', 'Reset the password via email'))
-            ->addSubcommand('resetAllPasswords')->setDescription(__d('CakeDC/Users', 'Reset the password for all users'))
-            ->addSubcommand('resetPassword')->setDescription(__d('CakeDC/Users', 'Reset the password for an specific user'))
+            ->addSubcommand('activateUser', [
+                'help' => __d('CakeDC/Users', 'Activate an specific user')
+            ])
+            ->addSubcommand('addSuperuser', [
+                'help' => __d('CakeDC/Users', 'Add a new superadmin user for testing purposes')
+            ])
+            ->addSubcommand('addUser', [
+                'help' => __d('CakeDC/Users', 'Add a new user')
+            ])
+            ->addSubcommand('changeRole', [
+                'help' => __d('CakeDC/Users', 'Change the role for an specific user')
+            ])
+            ->addSubcommand('deactivateUser', [
+                'help' => __d('CakeDC/Users', 'Deactivate an specific user')
+            ])
+            ->addSubcommand('deleteUser', [
+                'help' => __d('CakeDC/Users', 'Delete an specific user')
+            ])
+            ->addSubcommand('passwordEmail', [
+                'help' => __d('CakeDC/Users', 'Reset the password via email')
+            ])
+            ->addSubcommand('resetAllPasswords', [
+                'help' => __d('CakeDC/Users', 'Reset the password for all users')
+            ])
+            ->addSubcommand('resetPassword', [
+                'help' => __d('CakeDC/Users', 'Reset the password for an specific user')
+            ])
             ->addOptions([
                 'username' => ['short' => 'u', 'help' => 'The username for the new user'],
                 'password' => ['short' => 'p', 'help' => 'The password for the new user'],
@@ -112,7 +130,7 @@ class UsersShell extends Shell
     {
         $password = Hash::get($this->args, 0);
         if (empty($password)) {
-            $this->setError(__d('CakeDC/Users', 'Please enter a password.'));
+            $this->abort(__d('CakeDC/Users', 'Please enter a password.'));
         }
         $hashedPassword = $this->_generatedHashedPassword($password);
         $this->Users->updateAll(['password' => $hashedPassword], ['id IS NOT NULL']);
@@ -135,10 +153,10 @@ class UsersShell extends Shell
         $username = Hash::get($this->args, 0);
         $password = Hash::get($this->args, 1);
         if (empty($username)) {
-             $this->setError(__d('CakeDC/Users', 'Please enter a username.'));
+            $this->abort(__d('CakeDC/Users', 'Please enter a username.'));
         }
         if (empty($password)) {
-             $this->setError(__d('CakeDC/Users', 'Please enter a password.'));
+            $this->abort(__d('CakeDC/Users', 'Please enter a password.'));
         }
         $data = [
             'password' => $password
@@ -163,10 +181,10 @@ class UsersShell extends Shell
         $username = Hash::get($this->args, 0);
         $role = Hash::get($this->args, 1);
         if (empty($username)) {
-             $this->setError(__d('CakeDC/Users', 'Please enter a username.'));
+            $this->abort(__d('CakeDC/Users', 'Please enter a username.'));
         }
         if (empty($role)) {
-             $this->setError(__d('CakeDC/Users', 'Please enter a role.'));
+            $this->abort(__d('CakeDC/Users', 'Please enter a role.'));
         }
         $data = [
             'role' => $role
@@ -215,7 +233,7 @@ class UsersShell extends Shell
     {
         $reference = Hash::get($this->args, 0);
         if (empty($reference)) {
-             $this->setError(__d('CakeDC/Users', 'Please enter a username or email.'));
+            $this->abort(__d('CakeDC/Users', 'Please enter a username or email.'));
         }
         $resetUser = $this->Users->resetToken($reference, [
             'expiration' => Configure::read('Users.Token.expiration'),
@@ -227,7 +245,7 @@ class UsersShell extends Shell
             $this->out($msg);
         } else {
             $msg = __d('CakeDC/Users', 'The password token could not be generated. Please try again');
-             $this->setError($msg);
+            $this->abort($msg);
         }
     }
 
@@ -241,7 +259,7 @@ class UsersShell extends Shell
     {
         $username = Hash::get($this->args, 0);
         if (empty($username)) {
-             $this->setError(__d('CakeDC/Users', 'Please enter a username.'));
+            $this->abort(__d('CakeDC/Users', 'Please enter a username.'));
         }
         $data = [
             'active' => $active
@@ -316,7 +334,7 @@ class UsersShell extends Shell
     {
         $user = $this->Users->find()->where(['username' => $username])->first();
         if (empty($user)) {
-             $this->setError(__d('CakeDC/Users', 'The user was not found.'));
+            $this->abort(__d('CakeDC/Users', 'The user was not found.'));
         }
         $user = $this->Users->patchEntity($user, $data);
         collection($data)->filter(function ($value, $field) use ($user) {
@@ -338,7 +356,7 @@ class UsersShell extends Shell
     {
         $username = Hash::get($this->args, 0);
         if (empty($username)) {
-             $this->setError(__d('CakeDC/Users', 'Please enter a username.'));
+            $this->abort(__d('CakeDC/Users', 'Please enter a username.'));
         }
         $user = $this->Users->find()->where(['username' => $username])->first();
         if (isset($this->Users->SocialAccounts)) {
@@ -346,7 +364,7 @@ class UsersShell extends Shell
         }
         $deleteUser = $this->Users->delete($user);
         if (!$deleteUser) {
-             $this->setError(__d('CakeDC/Users', 'The user {0} was not deleted. Please try again', $username));
+            $this->abort(__d('CakeDC/Users', 'The user {0} was not deleted. Please try again', $username));
         }
         $this->out(__d('CakeDC/Users', 'The user {0} was deleted successfully', $username));
     }