+++ tests/TestCase/Controller/Component/UsersAuthComponentTest.php
@@ -19,6 +19,7 @@ use Cake\Core\Configure;
 use Cake\Core\Plugin;
 use Cake\Database\Exception;
 use Cake\Event\Event;
+use Cake\Http\ServerRequest;
 use Cake\Network\Request;
 use Cake\Network\Session;
 use Cake\ORM\Entity;
@@ -73,6 +74,7 @@ class UsersAuthComponentTest extends TestCase
             ->setMethods(['stop'])
             ->getMock();
         $this->Controller = new Controller($this->request, $this->response);
+        $this->Controller->name = 'Users';
         $this->Registry = $this->Controller->components();
         $this->Controller->UsersAuth = new UsersAuthComponent($this->Registry);
     }
@@ -266,7 +268,7 @@ class UsersAuthComponentTest extends TestCase
      * test
      *
      * @return void
-     * @expectedException Cake\Routing\Exception\MissingRouteException
+     * @expectedException \Cake\Routing\Exception\MissingRouteException
      */
     public function testIsUrlAuthorizedMissingRouteString()
     {
@@ -287,7 +289,7 @@ class UsersAuthComponentTest extends TestCase
      * test
      *
      * @return void
-     * @expectedException Cake\Routing\Exception\MissingRouteException
+     * @expectedException \Cake\Routing\Exception\MissingRouteException
      */
     public function testIsUrlAuthorizedMissingRouteArray()
     {
@@ -436,5 +438,84 @@ class UsersAuthComponentTest extends TestCase
             ->will($this->returnValue(true));
         $result = $this->Controller->UsersAuth->isUrlAuthorized($event);
         $this->assertTrue($result);
+    }
+
+    /**
+     * When application is installed using a base folder, we need to ensure array routes are
+     * normalized too to remove the base from the url used for matching the rules
+     *
+     * @see https://github.com/CakeDC/users/issues/538
+     *
+     * @return void
+     */
+    public function testIsUrlAuthorizedBaseUrl()
+    {
+        Configure::write('App.base', 'app');
+        Router::pushRequest(new ServerRequest([
+                'base' => '/app',
+                'url' => '/',
+            ]));
+        $event = new Event('event');
+        $event->data = [
+            'url' => [
+                'plugin' => 'CakeDC/Users',
+                'controller' => 'Users',
+                'action' => 'requestResetPassword',
+            ],
+        ];
+        $this->Controller->Auth = $this->getMockBuilder('Cake\Controller\Component\AuthComponent')
+            ->setMethods(['user', 'isAuthorized'])
+            ->disableOriginalConstructor()
+            ->getMock();
+        $this->Controller->Auth->expects($this->once())
+            ->method('user')
+            ->will($this->returnValue(['id' => 1]));
+        $this->Controller->Auth->expects($this->once())
+            ->method('isAuthorized')
+            ->with(null, $this->callback(function ($subject) {
+                return $subject->params === [
+                        'plugin' => 'CakeDC/Users',
+                        'controller' => 'Users',
+                        'action' => 'requestResetPassword',
+                        '_ext' => null,
+                        'pass' => [],
+                        '_matchedRoute' => '/route/*',
+                    ];
+            }))
+            ->will($this->returnValue(true));
+        $result = $this->Controller->UsersAuth->isUrlAuthorized($event);
+        $this->assertTrue($result);
+    }
+    /**
+     * test The user is logged in and allowed by rules to access this action,
+     * and we are checking another controller action not allowed
+     *
+     * this case would prevent permissions checked for allowed actions in another controller
+     * @see https://github.com/CakeDC/users/issues/527 for a workaround if you need to
+     * check allowed on another controller
+     *
+     * @return void
+     */
+    public function testIsUrlAuthorizedUserLoggedInAllowedActionAllowedAnotherController()
+    {
+        Router::connect('/route-another-controller/*', [
+            'plugin' => 'CakeDC/Users',
+            'controller' => 'AnotherController',
+            'action' => 'requestResetPassword'
+        ]);
+        $event = new Event('event');
+        $event->data = [
+            'url' => '/route-another-controller',
+        ];
+        $this->Controller->Auth = $this->getMockBuilder('Cake\Controller\Component\AuthComponent')
+            ->setMethods(['user', 'isAuthorized'])
+            ->disableOriginalConstructor()
+            ->getMock();
+        $this->Controller->Auth->allowedActions = ['requestResetPassword'];
+        $this->Controller->Auth->expects($this->once())
+            ->method('user')
+            ->will($this->returnValue(false));
+        $result = $this->Controller->UsersAuth->isUrlAuthorized($event);
+        $this->assertFalse($result);
     }
 }