+++ tests/TestCase/Model/Behavior/RegisterBehaviorTest.php
@@ -12,6 +12,7 @@
 
 namespace CakeDC\Users\Test\TestCase\Model\Behavior;
 
+use CakeDC\Users\Exception\UserAlreadyActiveException;
 use Cake\Core\Configure;
 use Cake\Mailer\Email;
 use Cake\ORM\TableRegistry;
@@ -314,5 +315,53 @@ class RegisterBehaviorTest extends TestCase
             'email_class' => $this->Email
         ]);
         $this->assertSame('emperor', $result['role']);
+    }
+
+    /**
+     * Test resendValidationEmail method
+     *
+     * @return void
+     */
+    public function testResendValidationEmail()
+    {
+        $user = [
+            'username' => 'testuser',
+            'email' => 'testuser@test.com',
+            'password' => 'password',
+            'password_confirm' => 'password',
+            'first_name' => 'test',
+            'last_name' => 'user',
+            'tos' => 1
+        ];
+        $result = $this->Table->register($this->Table->newEntity(), $user, ['token_expiration' => 3600, 'validate_email' => 1, 'email_class' => $this->Email]);
+        $this->assertFalse($result->active);
+        $originalExpiration = $result->token_expires;
+        $updatedResult = $this->Table->resendValidationEmail($result, ['token_expiration' => 4000, 'email_class' => $this->Email]);
+        $this->assertNotEmpty($updatedResult);
+        $this->assertFalse($updatedResult->active);
+        $newExpiration = $updatedResult->token_expires;
+        $this->assertNotEquals($originalExpiration, $newExpiration);
+    }
+
+    /**
+     * Test resendValidationEmail method throw exception on active user
+     *
+     * @return void
+     */
+    public function testResendValidationEmailThrows()
+    {
+        $user = [
+            'username' => 'testuser',
+            'email' => 'testuser@test.com',
+            'password' => 'password',
+            'password_confirm' => 'password',
+            'first_name' => 'test',
+            'last_name' => 'user',
+            'tos' => 1
+        ];
+        $result = $this->Table->register($this->Table->newEntity(), $user, ['token_expiration' => 3600, 'validate_email' => 1, 'email_class' => $this->Email]);
+        $activeUser = $this->Table->activateUser($result);
+        $this->expectException(UserAlreadyActiveException::class);
+        $updatedResult = $this->Table->resendValidationEmail($activeUser, ['token_expiration' => 4000, 'email_class' => $this->Email]);
     }
 }