+++ CHANGELOG.md
@@ -4,6 +4,18 @@ Changelog
 Releases for CakePHP 3
 -------------
 
+* 5.0.3
+  * Implemented event dispatching on social login
+  * Fixed bugs reported
+  * Don't check for allowed actions in other controllers
+
+* 5.0.2
+  * Fixed bug parsing rule urls when application installed in a subdirectory
+
+* 5.0.1
+  * Bugfix release
+  * Minor BR language improvements
+
 * 5.0.0
   * Some Auth objects refactored into https://github.com/CakeDC/auth
   * Upgrade to CakePHP 3.4